Adjust directory structure and naming
