+++ apps/ecommerce/frontend/src/components/App/Dialog/ResourceCard/ResourceCard.styles.ts
