+++ apps/ecommerce/frontend/src/components/App/Field/ResourceCard/ResourceCardRawData/ResourceCardRawData.tsx
