+++ apps/ecommerce/frontend/src/components/App/Field/ResourceCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ResourceCard/ResourceCard';
