+++ apps/ecommerce/frontend/src/components/App/Field/ResourceField/index.ts
@@ -1,0 +1,2 @@
+export { default as ResourceField } from './ResourceField/ResourceField';
+export { default as FieldFallback } from './FieldFallback/FieldFallback';
