+++ apps/ecommerce/frontend/src/components/App/Field/index.ts
@@ -1,0 +1,3 @@
+export { default as SingleResource } from './SingleResource/SingleResource';
+export { default as MultipleResource } from './MultipleResources/MultipleResources';
+export { FieldFallback } from './ResourceField';
