+++ apps/ecommerce/frontend/src/components/Config/AppLogo/AppLogo.spec.tsx
