+++ apps/ecommerce/frontend/src/components/Config/ConfigPage/ConfigPage.spec.tsx
@@ -1,0 +1,16 @@
+import { render, screen } from '@testing-library/react';
+import { mockSdk, mockCma } from '../../../../test/mocks';
+import ConfigPage from './ConfigPage';
+
+jest.mock('@contentful/react-apps-toolkit', () => ({
+  useSDK: () => mockSdk,
+  useCMA: () => mockCma,
+}));
+
+describe('Config Page component', () => {
+  it('mounts', () => {
+    render(<ConfigPage />);
+
+    expect(screen.getByTestId('configPageBackground')).toBeInTheDocument();
+  });
+});
