+++ apps/ecommerce/frontend/src/components/Config/ErrorDisplay/ErrorDisplay.tsx
