--- apps/ecommerce/frontend/src/components/Field/ResourceCard/ResourceCardMenu.tsx
@@ -1,29 +1,0 @@
-import { IconButton, Menu } from '@contentful/f36-components';
-import { MoreHorizontalIcon } from '@contentful/f36-icons';
-import useResourceMenu from 'components/Field/ResourceCard/useResourceMenuItems';
-
-export interface ResourceCardMenuProps {
-  onRemove: Function;
-  isDataVisible: boolean;
-  onShowData: Function;
-  onHideData: Function;
-  index?: number;
-  total?: number;
-  onMoveToTop?: Function;
-  onMoveToBottom?: Function;
-}
-
-const ResourceCardMenu = (props: ResourceCardMenuProps) => {
-  const { menuItems } = useResourceMenu(props);
-
-  return (
-    <Menu offset={[-5, 0]}>
-      <Menu.Trigger>
-        <IconButton icon={<MoreHorizontalIcon />} aria-label="Actions" />
-      </Menu.Trigger>
-      <Menu.List>{menuItems}</Menu.List>
-    </Menu>
-  );
-};
-
-export default ResourceCardMenu;
