--- apps/ecommerce/frontend/src/components/Field/ResourceCard/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ResourceCard';
