--- apps/ecommerce/frontend/src/components/Field/ResourceField/index.ts
@@ -1,2 +1,0 @@
-export { default as ResourceField } from './ResourceField';
-export { default as FieldFallback } from './FieldFallback';
