--- apps/ecommerce/frontend/src/components/Field/Wrappers/index.ts
@@ -1,2 +1,0 @@
-export { default as MultipleResources } from './MultipleResources';
-export { default as SingleResource } from './SingleResource';
