--- apps/ecommerce/frontend/src/components/Field/index.ts
@@ -1,2 +1,0 @@
-export { SingleResource, MultipleResources } from './Wrappers';
-export { FieldFallback } from './ResourceField';
