+++ apps/ecommerce/frontend/src/locations/Dialog.tsx
@@ -1,4 +1,4 @@
-import ResourceSelectionDialog from 'components/Dialog/ResourceSelectionDialog';
+import ResourceSelectionDialog from 'components/Dialog/ResourceSelectionDialog/ResourceSelectionDialog';
 
 const Dialog = () => {
   return <ResourceSelectionDialog />;
